+++ src/exchanges/poloniex.js
@@ -107,6 +107,30 @@ class Poloniex {
     })
   }
 
+  address(asset) {
+    return new Promise((resolve, reject) => {
+      this.plnx.returnDepositAddresses()
+        .then( addresses => {
+          let address = addresses[asset]
+
+          if(address) {
+            resolve(address)
+          } else {
+            this.plnx.generateNewAddress(asset)
+              .then( response => {
+                if(response.success) {
+                  resolve(response.response)
+                } else {
+                  reject('Could not create address.')
+                }
+              })
+              .catch(err => reject(err.message))
+          }
+        })
+        .catch(err => reject(err.message))
+    })
+  }
+
 }
 
 module.exports = Poloniex