+++ src/exchanges/yunbi.js
@@ -129,6 +129,13 @@ class Yunbi {
   //   })
   // }
 
+  address(asset) {
+    return new Promise((resolve, reject) => {
+      // TODO: fetch addresses
+      reject('Not implemented.')
+    })
+  }
+
 }
 
 module.exports = Yunbi